avcodec: Add an option for trimming out the audio encoder preroll from the content itself
This avoids extending the encoded audio, while compromising the
content of the start of the audio stream.

This allows e.g. encoding audio without shifting data forwards, for
containers that lack signalling of the preroll (or for e.g. avoiding
edit lists in mp4, to ease conformance with the DASH-IF interoperability
guidelines).

This is only implemented in avcodec_encode_audio2 right now, not
in avcodec_send_frame (no encoders currently implement the
send_frame function).