movenc: switch the ISMV tfxd fragment timestamp to be DTS
As far as can be seen, this value is supposed to be the DTS of a
fragment in smooth streaming. Thus, don't take the initial sample's
CTS offset into mention when calculating the start timestamp.

This can be especially seen with H.264 with b-frames in smooth
streaming, where implementations will first parse the txfd
timestamp, and then once more add the CTS offset of the first
sample to that timestamp to get the presentation time of that
sample, leading to even more A/V desync and/or actual
discontinuities when the DTS+(2*CTS offset)+duration value
ends up larger than the following tfxd's start timestamp.

The duration calculation requires PTS values, so it is not touched.