movenc: utilize negative CTS offsets by default with ISMV
Reasons for change:

* ISMV has no edit lists.
* As the CTS offset of the initial sample is now zero, we can more
  easily synchronize the time lines between tracks' fragments.

  With positive CTS offsets the current code does not try to
  adjust the start times of the ISMV fragments and thus in other
  words, not using negative CTS offsets with ISMV just
  doesn't seem to be like a working solution as soon as something
  like B-frames come into play.

Tests have been updated and the diff has been verified with
L-SMASH's boxdumper that the `tfxd` values are now zero at the
beginning, and that the initial CTS offset of the first video
sample is zero.

The first video sample gets its proper single sample
duration added on top of the initial start offset of the track
pushed into the initial sample's duration. The additional, now
unneeded, duration will get handled in the following commit.